requirements.txt
google-auth
requests
flask
playwright